feat: extract brush undo cloning helper
