+++ Assets/Scripts/App.cs
@@ -16,29 +16,12 @@ using UnityEngine;
 
 namespace TiltBrush
 {
-    public class App : MonoBehaviour
+    public static class App
     {
-        public const float METERS_TO_UNITS = 10f;
-        public const float UNITS_TO_METERS = .1f;
+        public const float METERS_TO_UNITS = AppCore.METERS_TO_UNITS;
+        public const float UNITS_TO_METERS = AppCore.UNITS_TO_METERS;
 
-        /// Time origin of sketch in seconds for case when drawing is not sync'd to media.
-        private static double m_sketchTimeBase = 0;
         public static double CurrentSketchTime =>
-            // Unity's Time.time has useful precision probably <= 1ms, and unknown
-            // drift/accuracy. It is a single (but is a double, internally), so its
-            // raw precision drops to ~2ms after ~4 hours and so on.
-            // Time.timeSinceLevelLoad is also an option.
-            //
-            // C#'s DateTime API has low-ish precision (10+ ms depending on OS)
-            // but likely the highest accuracy with respect to wallclock, since
-            // it's reading from an RTC.
-            //
-            // High-precision timers are the opposite: high precision, but are
-            // subject to drift.
-            //
-            // For realtime sync, Time.time is probably the best thing to use.
-            // For postproduction sync, probably C# DateTime.
-            // If you change this, also modify SketchTimeToLevelLoadTime
-            Time.timeSinceLevelLoad - m_sketchTimeBase;
+            AppCore.GetCurrentSketchTime(Time.timeSinceLevelLoad);
     }
 }
