+++ Assets/Scripts/Canvas.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c474728e39a34b06ac55d0f7d4c01d43
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
